requirements.txt
pyyaml
python-telegram-bot
python-bugzilla
pymongo